Change the way you add query parameters Haulmont/jmix-reports#120
If the entities are selected by query, QueryStepFragment appears as the third step. The initial query that is shown in QueryStepFragment is generated by JpqlQueryBuilder depending on the properties selected in the second step.
If there are parameters in the query it is possible to generate report parameters but it is required to set the actual type for them. Additionally, it is possible to add a default value for the parameter (except a list of entities).
The query is regenerated and created parameters are reset if the new region is added or properties are changed.